backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
python-jose==3.3.0
passlib==1.7.4
requests==2.31.0
cryptography==41.0.7
